manhattan {qqman}	R Documentation

Creates a manhattan plot

Description

Creates a manhattan plot from PLINK assoc output (or any data frame with chromosome, position, and p-value).

Usage

manhattan(
  x,
  chr = "CHR",
  bp = "BP",
  p = "P",
  snp = "SNP",
  col = c("gray10", "gray60"),
  chrlabs = NULL,
  suggestiveline = -log10(1e-05),
  genomewideline = -log10(5e-08),
  highlight = NULL,
  logp = TRUE,
  annotatePval = NULL,
  annotateTop = TRUE,
  ...
)

Arguments

x
A data.frame with columns "BP," "CHR," "P," and optionally, "SNP."

chr
A string denoting the column name for the chromosome. Defaults to PLINK's "CHR." Said column must be numeric. If you have X, Y, or MT chromosomes, be sure to renumber these 23, 24, 25, etc.

bp
A string denoting the column name for the chromosomal position. Defaults to PLINK's "BP." Said column must be numeric.

p
A string denoting the column name for the p-value. Defaults to PLINK's "P." Said column must be numeric.

snp
A string denoting the column name for the SNP name (rs number). Defaults to PLINK's "SNP." Said column should be a character.

col
A character vector indicating which colors to alternate.

chrlabs
A character vector equal to the number of chromosomes specifying the chromosome labels (e.g., c(1:22, "X", "Y", "MT")).

suggestiveline
Where to draw a "suggestive" line. Default -log10(1e-5). Set to FALSE to disable.

genomewideline
Where to draw a "genome-wide sigificant" line. Default -log10(5e-8). Set to FALSE to disable.

highlight
A character vector of SNPs in your dataset to highlight. These SNPs should all be in your dataset.

logp
If TRUE, the -log10 of the p-value is plotted. It isn't very useful to plot raw p-values, but plotting the raw value could be useful for other genome-wide plots, for example, peak heights, bayes factors, test statistics, other "scores," etc.

annotatePval
If set, SNPs below this p-value will be annotated on the plot. If logp is FALSE, SNPs above the specified value will be annotated.

annotateTop
If TRUE, only annotates the top hit on each chromosome that is below the annotatePval threshold (or above if logp is FALSE).

...
Arguments passed on to other plot/points functions

Value

A manhattan plot.

Examples

manhattan(gwasResults)
  

[Package qqman version 0.1.9 Index]